Create UBO_Minimal_Porn_Lingerie_MyFiltersList_inc_whitelisted_Items
This minimal list uses $all blocking for keywords and then selective @@ whitelisted sites. The idea is to keep porn and lingerie images and pages from browsing. keyword doesn't get all of the adult sides but you can add and tweak as you like or look at Site Blocker addon which has a pornography category but this does the job in filtering URLs for the most part.